tools: zynqmpimage: Add bif support
The officially described way to generate boot.bin files for ZynqMP is to
describe the contents of the target binary using a file of the "bif"
format.  This file then links to other files that all get packed into a
bootable image.

This patch adds support to read such a .bif file and generate a respective
ZynqMP boot.bin file that can include the normal image and pmu files, but
also supports image partitions now. This makes it a handy replacement for
the proprietary "bootgen" utility that is currently used to generate
boot.bin files with FSBL.

Signed-off-by: Alexander Graf <agraf@suse.de>

---

v2 -> v3:

  - zero initialize header
  - reduce default debug verbosity

v3 -> v4:

  - add error handling
  - add fsbl_config support
  - add aarch32 support
  - allow a5x to be written as a53
  - add offset support
  - add support for partition_owner
  - ensure pmufw comes before bootloader
  - simplify fsbl_config
  - add non-a53 boot support
  - checkpatch fixes